requirements.txt
# Requirements for query_llms.py
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0

# Requirements for image_generation.py
requests>=2.31.0
replicate>=0.20.0

